+++ packages/wallet-connectors/CHANGELOG.md
@@ -9,7 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
--   `WalletConnection`: Support both module and parameter schemas in `signAndSendTransaction`.
+-   `WalletConnection` (breaking): Support both module and parameter schemas in `signAndSendTransaction`.
+    To migrate existing usage, wrap the schema string in the new function `moduleSchema(...)`.
 
 ## [0.2.3] - 2023-04-03